novo codigo

+++ Salario.py
@@ -1,4 +1,4 @@
 nome = input("DIgite seu nome : ")
-idade= float(input("Digite sua idade: "))
+idade= int(input("Digite sua idade: "))
 salario = float(input("Digite seu salario: "))
 print(f'Olá {nome}, você tem {idade}, e seu salario é {salario}')
